services/imageresizer/requirements.txt
Flask==1.1.2
google-cloud-storage==1.35.0
Pillow==8.1.0